refactor leverage slider
